移動とかジャンプとかの動作を正常化できた。摩擦による減速が空気中でも変わらないのが気になる

+++ Aim.h
@@ -13,8 +13,8 @@ class Aim : public GameObject
     Player* pPlayer_;                       //プレイヤーインスタンス
     XMFLOAT3 aimDirection_;                 //現在の視点に基づいた進行方向ベクトル
     XMFLOAT2 mouseSensitivity_;             //マウスの感度
-    const float MOUSE_SENSITIVITY_X = 0.05;
-    const float MOUSE_SENSITIVITY_Y = 0.05;
+    const double MOUSE_SENSITIVITY_X = 0.05;
+    const double MOUSE_SENSITIVITY_Y = 0.05;
 
 public:
     Aim(GameObject* parent);        //コンストラクタ
